Move validation to helpers
